pyproject.toml
[tool.poetry]
name = "slack-bot"
version = "0.1.0"
description = ""
authors = ["furuichi-kotaro <54981611+furuich-kotaro@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "slack_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
aws-wsgi = "^0.2.7"
slack-bolt = "^1.18.0"
langchain = "^0.0.173"
openai = "^0.27.6"
requests = "2.28.1"
pinecone-client = "^2.2.1"
pinecone-text = "^0.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
